Module Quantumlib

module C_emitter : sig ... end

C Code emission. This module implements C code generation for quantum circuits (see Circuit.t). It uses a symbolic version of the simulator (see Symbolic and Simulation).

module Circuit : sig ... end

A simplified representation of Quantum Circuits.

module Compiler : sig ... end

A compiler from Qasm to circuits.

module Example : sig ... end

A collection of examples using circuits expressed in Circuit.t

module Naive : sig ... end

A naive simulator for quantum circuits

module Parser : sig ... end

OpenQASM parser (supports OpenQASM 2.0)

module Qasm : sig ... end

Abstract Syntax Tree of OpenQASM

module Se : sig ... end

A symbolic simulator for quantum circuits

module Simulation : sig ... end

Interface of simulators for quantum circuits

module Symbolic : sig ... end

Symbolic computations. This module implements symbolic computations for complex-valued expressions. It is used by the symbolic simulator Se.

module Utils : sig ... end