Module Quantumlib.C_emitter

C Code emission. This module implements C code generation for quantum circuits (see Circuit.t). It uses a symbolic version of the simulator (see Symbolic and Simulation).

val emitc : Circuit.t -> Stdlib.out_channel -> unit

emitc circ out is generating C code for the simulation of circuit circ in the given output channel out.