Module Quantumlib.Circuit

A simplified representation of Quantum Circuits.

type adr =

| A of int

type arb_sq_gate = {

theta : float;

phi : float;

lambda : float;

}

An arbitrary single-qubit gate, parametrized as in OpenQASM

type gate_kind =

| X

| H

| Z

| U of arb_sq_gate

| Rm of int

Kind of Gates.

type gate = {

target : adr;

kind : gate_kind;

controls : adr list;

}

Generic Gates. A gate has a target qubit and a set of set of controls.

type t = {

qbits : int;

gates : gate list;

}

Circuits. A quantum circuit is defined by its number a qubits and an (ordered) list of gates.